+++ src/main/java/net/yxiao233/ifeu/IndustrialForegoingExtraUpgrades.java
@@ -24,6 +24,8 @@ public class IndustrialForegoingExtraUpgrades extends ModuleController {
 
         IFEUContents.BLOCKS.register(modEventBus);
         IFEUContents.ITEMS.register(modEventBus);
+        IFEUContents.TRUNK_PLACER_TYPES.register(modEventBus);
+        IFEUContents.FOLIAGE_PLACER_TYPES.register(modEventBus);
         IFEUCreativeModeTab.CREATIVE_MODE_TAB.register(modEventBus);
         IFEUDataComponentTypes.DATA_COMPONENTS.register(modEventBus);
     }