+++ src/main/java/net/yxiao233/ifeu/client/data/ModBlockTagProvider.java
@@ -2,8 +2,12 @@ package net.yxiao233.ifeu.client.data;
 
 import com.buuz135.industrial.module.ModuleCore;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.data.PackOutput;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.world.level.block.Block;
 import net.neoforged.neoforge.common.Tags;
 import net.neoforged.neoforge.common.data.BlockTagsProvider;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
@@ -14,9 +18,12 @@ import net.yxiao233.ifeu.common.registry.IFEUContents;
 import net.yxiao233.ifeu.common.registry.IFEUTags;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 public class ModBlockTagProvider extends BlockTagsProvider {
+    private final List<ResourceLocation> generated = new ArrayList<>();
     public ModBlockTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, @Nullable ExistingFileHelper existingFileHelper) {
         super(output, lookupProvider, IndustrialForegoingExtraUpgrades.MODID, existingFileHelper);
     }
@@ -66,11 +73,22 @@ public class ModBlockTagProvider extends BlockTagsProvider {
                 .add(IFEUBlocks.FERMENTER.getBlock());
 
         DeferredTree.DeferredTreesRegister.getAllTrees().forEach(tree ->{
-            this.tag(BlockTags.LOGS_THAT_BURN)
-                    .add(tree.getLogBlock().get())
-                    .add(tree.getStrippedLogBlock().get())
-                    .add(tree.getWoodBlock().get())
-                    .add(tree.getStrippedWoodBlock().get());
+            if(shouldGenerated(tree.getLogBlock().get())){
+                this.tag(BlockTags.LOGS_THAT_BURN).add(tree.getLogBlock().get());
+                generated(tree.getLogBlock().get());
+            }
+            if(shouldGenerated(tree.getStrippedLogBlock().get())){
+                this.tag(BlockTags.LOGS_THAT_BURN).add(tree.getStrippedLogBlock().get());
+                generated(tree.getStrippedLogBlock().get());
+            }
+            if(shouldGenerated(tree.getWoodBlock().get())){
+                this.tag(BlockTags.LOGS_THAT_BURN).add(tree.getWoodBlock().get());
+                generated(tree.getWoodBlock().get());
+            }
+            if(shouldGenerated(tree.getStrippedWoodBlock().get())){
+                this.tag(BlockTags.LOGS_THAT_BURN).add(tree.getStrippedWoodBlock().get());
+                generated(tree.getStrippedWoodBlock().get());
+            }
         });
 
 
@@ -103,5 +121,14 @@ public class ModBlockTagProvider extends BlockTagsProvider {
         this.tag(IFEUTags.Blocks.STORAGE_ENERGY)
 //                .addTag(IFEUTags.Blocks.BLACK_HOLE_CAPACITOR)
                 .add(IFEUBlocks.CREATIVE_CAPACITOR.getBlock());
+    }
+
+    private boolean shouldGenerated(Block block){
+        ResourceLocation location = BuiltInRegistries.BLOCK.getKey(block);
+        return !generated.contains(location);
+    }
+
+    private void generated(Block block){
+        generated.add(BuiltInRegistries.BLOCK.getKey(block));
     }
 }